Split out value types from instance.go
