+++ policy/model/instance.go
@@ -15,10 +15,6 @@
 // Package model contains abstract representations of policy template and instance objects.
 package model
 
-import (
-	structpb "github.com/golang/protobuf/ptypes/struct"
-)
-
 // Instance declares the properties common to all policy instances.
 //
 // Each instance must specify its version and kind, as well as its name in the metadata field.
@@ -106,76 +102,3 @@ type ExprMatcher struct {
 	Operator *DynValue
 	Values   *DynValue
 }
-
-// DynValue is a dynamically typed value used to describe unstructured content.
-// Whether the value has the desired type is determined by where it is used within the Instance or
-// Template, and whether there are schemas which might enforce a more rigid type definition.
-type DynValue struct {
-	ID    int64
-	Value ValueNode
-}
-
-// ValueNode is a marker interface used to indicate which value types may populate a DynValue's
-// Value field.
-type ValueNode interface {
-	isValueNode()
-}
-
-// StructValue declares an object with a set of named fields whose values are dynamically typed.
-type StructValue struct {
-	Fields []*StructField
-}
-
-func (*StructValue) isValueNode() {}
-
-// StructField specifies a field name and a reference to a dynamic value.
-type StructField struct {
-	ID   int64
-	Name string
-	Ref  *DynValue
-}
-
-// ListValue contains a list of dynamically typed entries.
-type ListValue struct {
-	Entries []*DynValue
-}
-
-func (*ListValue) isValueNode() {}
-
-// BoolValue is a boolean value suitable for use within DynValue objects.
-type BoolValue bool
-
-func (BoolValue) isValueNode() {}
-
-// BytesValue is a []byte value suitable for use within DynValue objects.
-type BytesValue []byte
-
-func (BytesValue) isValueNode() {}
-
-// DoubleValue is a float64 value suitable for use within DynValue objects.
-type DoubleValue float64
-
-func (DoubleValue) isValueNode() {}
-
-// IntValue is an int64 value suitable for use within DynValue objects.
-type IntValue int64
-
-func (IntValue) isValueNode() {}
-
-// NullValue is a protobuf.Struct concrete null value suitable for use within DynValue objects.
-type NullValue structpb.NullValue
-
-func (NullValue) isValueNode() {}
-
-// StringValue is a string value suitable for use within DynValue objects.
-type StringValue string
-
-func (StringValue) isValueNode() {}
-
-// UintValue is a uint64 value suitable for use within DynValue objects.
-type UintValue uint64
-
-func (UintValue) isValueNode() {}
-
-// Null is a singleton NullValue instance.
-var Null NullValue
